+++ src/Cli.php
@@ -8,6 +8,7 @@
  * PHP version 8.1
  *
  * @package Inane\Cli
+ * @category console
  *
  * @author    	James Logsdon <dwarf@girsbrain.org>
  * @author		Philip Michael Raab<peep@inane.co.za>
@@ -84,6 +85,15 @@ class Cli {
     }
 
     /**
+     * Plays a beep sound
+     *
+     * @return void
+     */
+    public static function beep(): void {
+        echo "\x07";
+    }
+
+    /**
      * Handles rendering strings. If extra scalar arguments are given after the `$msg`
      * the string will be rendered with `sprintf`. If the second argument is an `array`
      * then each key in the array will be the placeholder name. Placeholders are of the
@@ -91,6 +101,7 @@ class Cli {
      *
      * @param string   $msg  The message to render.
      * @param mixed    ...   Either scalar arguments or a single array argument.
+     *
      * @return string  The rendered string.
      */
     public static function render($msg) {
@@ -103,7 +114,9 @@ class Cli {
      *
      * @param string  $msg  The message to output in `printf` format.
      * @param mixed   ...   Either scalar arguments or a single array argument.
+     *
      * @return void
+     *
      * @see \Inane\Cli\render()
      */
     public static function out(string $msg) {
@@ -473,18 +486,13 @@ class Cli {
     public static function getUnicodeRegexs($idx = null) {
         static $eaw_regex; // East Asian Width regex. Characters that count as 2 characters as they're "wide" or "fullwidth". See http://www.unicode.org/reports/tr11/tr11-19.html
         static $m_regex; // Mark characters regex (Unicode property "M") - mark combining "Mc", mark enclosing "Me" and mark non-spacing "Mn" chars that should be ignored for spacing purposes.
-        if (null === $eaw_regex) {
-            // Load both regexs generated from Unicode data.
-            require __DIR__ . '/unicode/regex.php';
-        }
+
+        // Load both regexs generated from Unicode data.
+        if (null === $eaw_regex) require __DIR__ . '/unicode/regex.php';
 
         if (null !== $idx) {
-            if ('eaw' === $idx) {
-                return $eaw_regex;
-            }
-            if ('m' === $idx) {
-                return $m_regex;
-            }
+            if ('eaw' === $idx) return $eaw_regex;
+            if ('m' === $idx) return $m_regex;
         }
 
         return [$eaw_regex, $m_regex,];