+++ src/Pencil.php
@@ -8,6 +8,7 @@
  * PHP version 8.1
  *
  * @package Inane\Cli
+ * @category console
  *
  * @author		Philip Michael Raab<peep@inane.co.za>
  *
@@ -26,6 +27,7 @@ use Stringable;
 
 use function array_key_exists;
 use function fwrite;
+use function in_array;
 use function is_null;
 use function strlen;
 use const null;
@@ -45,7 +47,7 @@ use Inane\Cli\Pencil\{
  *
  * @package Inane\Cli
  *
- * @version 0.2.0
+ * @version 0.3.0
  */
 class Pencil implements Stringable {
     public const VERSION = '0.1.0';
@@ -62,6 +64,9 @@ class Pencil implements Stringable {
     /**
      * Pencil constructor
      *
+     * If $colour not set, the current colour and style remains in effect.
+     * $style only takes effect if a colour is set.
+     *
      * @param \Inane\Cli\Pencil\Colour $colour
      * @param \Inane\Cli\Pencil\Style $style
      * @param null|\Inane\Cli\Pencil\Colour $background
@@ -74,13 +79,15 @@ class Pencil implements Stringable {
          *
          * @var \Inane\Cli\Pencil\Colour
          */
-        private Colour $colour = Colour::Black,
+        private ?Colour $colour = null,
         /**
          * Pencil style
          *
+         * Only applicable if $colour is set.
+         *
          * @var \Inane\Cli\Pencil\Style
          */
-        private Style $style = Style::Regular,
+        private Style $style = Style::Plain,
         /**
          * Pencil background colour
          *
@@ -88,35 +95,53 @@ class Pencil implements Stringable {
          */
         private ?Colour $background = null,
     ) {
+        if ($style == Style::Hidden && in_array($colour, [Colour::Default, null])) $this->colour = Colour::Black;
     }
 
-    public static function reverse(string $text): ?string {
+    /**
+     * Returns the original uncoloured text from cache
+     *
+     * Useful for functions like `strlen` which would return values no consistent
+     *  with the actual size on screen.
+     *
+     * @param string $text coloured text
+     *
+     * @return null|string uncoloured text
+     */
+    public static function original(string $text): ?string {
         if (array_key_exists($text, static::$cache))
             return static::$cache[$text];
 
         return null;
     }
 
+    /**
+     * Returns the width of the original non-coloured text
+     *
+     * @param string $text coloured text
+     *
+     * @return null|int width of non-coloured text
+     */
     public static function width(string $text): ?int {
-        $string = static::reverse($text);
+        $string = static::original($text);
         if (!is_null($string)) return strlen($string);
 
         return null;
     }
 
-        /**
-	 * Pad the string to a certain display length.
-	 *
-	 * @param string      $string         The string to pad.
-	 * @param int         $length         The display length.
-	 * @param bool        $pre_colourised Optional. Set if the string is pre-colourised. Default false.
-	 * @param string|bool $encoding       Optional. The encoding of the string. Default false.
-	 * @param int         $pad_type       Optional. Can be STR_PAD_RIGHT, STR_PAD_LEFT, or STR_PAD_BOTH. If pad_type is not specified it is assumed to be STR_PAD_RIGHT.
-	 *
-	 * @return string
-	 */
+    /**
+     * Pad the string to a certain display length.
+     *
+     * @param string      $string         The string to pad.
+     * @param int         $length         The display length.
+     * @param bool        $pre_colourised Optional. Set if the string is pre-colourised. Default false.
+     * @param string|bool $encoding       Optional. The encoding of the string. Default false.
+     * @param int         $pad_type       Optional. Can be STR_PAD_RIGHT, STR_PAD_LEFT, or STR_PAD_BOTH. If pad_type is not specified it is assumed to be STR_PAD_RIGHT.
+     *
+     * @return string
+     */
     public static function pad(string $text, int $length, bool $pre_colourised = false, bool|string $encoding = false, int $pad_type = STR_PAD_RIGHT): ?string {
-        $string = static::reverse($text);
+        $string = static::original($text);
         if (!is_null($string)) {
             $real_length = static::width($text);
             $diff = strlen($text) - $real_length;
@@ -142,13 +167,19 @@ class Pencil implements Stringable {
      *
      * Value stored in <strong>Pencil::pencil</strong> af first calculation.
      *
+     * @param bool $reset when modified
+     *
      * @return string
      */
-    protected function getPencil(): string {
-        if (!isset($this->pencil)) {
-            $colour = Type::Plain->value + $this->colour->value;
-            $pencil = "\033[{$this->style->value};{$colour}m";
+    protected function getPencil(bool $reset = false): string {
+        if (!isset($this->pencil) || $reset) {
+            $pencil = '';
+            if (!is_null($this->colour)) {
+                $colour = $this->colour->value >= 0 ? Type::Plain->value + $this->colour->value : 0;
 
+                $style = is_null($this->style) ? '' : "{$this->style->value};";
+                $pencil .= "\033[{$style}{$colour}m";
+            }
             if (!is_null($this->background)) {
                 $background = Type::Highlight->value + $this->background->value;
                 $pencil .= "\033[{$background}m";
@@ -190,7 +221,31 @@ class Pencil implements Stringable {
      */
     public function line(string $text, bool $reset = true): void {
         $this->out("$text", $reset);
-        $this->out("\n", false);
+        fwrite(STDOUT, "\n");
+    }
+
+    /**
+     * get input from terminal
+     *
+     * Takes input from `STDIN` in the given format. If an end of transmission
+     * character is sent (^D), an exception is thrown.
+     *
+     * @since 0.3.0
+     *
+     * @param null|string	$format		A valid input format. See `fscanf`. If null all input to first newline as string.
+     * @param mixed			$default	Value to return if not an interactive terminal.
+     * @param bool          $reset      reset colours after reading
+     *
+     * @return mixed
+     *
+     * @throws \Exception
+     */
+    public function input(?string $format = null, mixed $default = null, bool $reset = true): mixed {
+        $this->out('', false);
+        $input = Cli::input($format, $default);
+        $this->out('', $reset);
+
+        return $input;
     }
 
     /**